+++ lib/runtime-abi/src/lib.rs
@@ -1,3 +1,5 @@
+#![deny(unused_imports, unused_variables, unused_unsafe, unreachable_patterns)]
+
 #[cfg(not(target_os = "windows"))]
 #[macro_use]
 extern crate failure;