+++ lib/singlepass-backend/src/lib.rs
@@ -1,3 +1,4 @@
+#![deny(unused_imports, unused_variables, unused_unsafe, unreachable_patterns)]
 #![feature(proc_macro_hygiene)]
 
 #[cfg(not(any(