+++ src/bin/wasmer.rs
@@ -1,3 +1,5 @@
+#![deny(unused_imports, unused_variables, unused_unsafe, unreachable_patterns)]
+
 extern crate structopt;
 
 use std::env;
@@ -86,6 +88,10 @@ struct Run {
     /// Emscripten symbol map
     #[structopt(long = "em-symbol-map", parse(from_os_str), group = "emscripten")]
     em_symbol_map: Option<PathBuf>,
+
+    /// Begin execution at the specified symbol
+    #[structopt(long = "em-entrypoint", group = "emscripten")]
+    em_entrypoint: Option<String>,
 
     /// WASI pre-opened directory
     #[structopt(long = "dir", multiple = true, group = "wasi")]
@@ -394,6 +400,7 @@ fn execute_wasm(options: &Run) -> Result<(), String> {
                 options.path.to_str().unwrap()
             },
             options.args.iter().map(|arg| arg.as_str()).collect(),
+            options.em_entrypoint.clone(),
         )
         .map_err(|e| format!("{:?}", e))?;
     } else {